create_node() Broken
Due to how Node.__init__() works, create_node() is broken.
In order to reduce the number of db.session.commit()s,
the hierarchy is assigned to a node while ambiguous.

This apparently doesn't play well with trying to add a node to an existing hierarchy.
Next steps are going back to simple_test.py to see what's up.
Current tree is a bit too complex to try and parse it out.