Value Functions, Weight Rebalancing, Direct Assessment (#84)
* Created balance_weights()

A generalized function to perform the calculation for
Direct Assessment. I still need a way to change weights.

* app.py rank_alternatives()

May have cracked the algorithm to rank alternatives.

Currently returns an unsorted dict of alternatives with calculated values.

Just kidding, the values aren't calculated because value functions aren't implemented.

* node.py normalize()

Basic linear equations should work... hopefully

* Changed weight to local_weight and global_weight

More accurately reflects the hiearchy model.

Global weight is currently calculated by multiplying
the local_weight by the parent's global_weight

* Set root local and global weights to 1

* rank_alternatives route was wrong# Please enter the commit message for your changes. Lines starting

* Initial Rankings seems to work

Now to test linear value functions

* Fixed Root weighting, changed checks in normalize()

* Made normalize()'s Linear section more normal

* Now can add alternatives with just a name

* Added comments from meeting with Luke

* Value Function Refreshing

Global values are automatically stored and updated
so that the frontend may get the list of rankings by
using the returned alternatives

* Cleaned up app.py

* Automatic Weight Rebalancing

Weights rebalance on hierarchy and node creation.

* Weights/Values refresh on node deletion. Rounding.

* Added vf_type field to node.py

* localValue and globalValue fields fixed

* Initial Reference Table

Switched to storing linear points on the table Reference

* Initial patch_node()

Not sure if this is what it should be.
Need to figure out when the front end will use it

* Initial direct_assessment()

I think this should work. I need to build the route on postman to double check

* direct_assessment() works

* Forgot to refresh alternatives

* Default local weight is 1

* valueFunctionData field returns 11 points for value function display

* TODO: Move valueFunctionData outside, save it, base off range

* valueFunctionData now is based off the range

* removed TODO

* Fixed swapping sections

* simplified valueFunctionData calculation

* Rounding only occurs on to_dict()

Made it so that y values more than 1 are 1
Less than 0 are 0

* patch_node() route works

* Fixed refresh_weights() error involving root in patch_node()

* Fixed list index out of range

* Linear functions create 2 reference points by default

Should fix some of the errors Luke's seeing

* value.py to_dict() checks that local_values are Floats

* Fixed dumb mistake

* Fixed if statement in node.py create()

* Fixed indexing error

* Fixed unnecessary break

* Convert measurement to Linear

Need to create origin points or given points

* patch_node() should be fixed

* direct_assessment new data format

the data object is now an array of objects

* Fixed conditional

* direct_assessment() works for new data format

* delete_node() returns parent of deleted node

* Simplified normalize(), fixed issue with decreasing functions in to_dict()

* Fixed issue with value.py to_dict()